@@ -1,5 +1,8 @@
 <?php
 
+use App\Http\Controllers\PaymentController;
+use Illuminate\Support\Facades\Route;
+
 /*
 |--------------------------------------------------------------------------
 | Web Routes
@@ -15,8 +18,6 @@ Route::get('/', function () {
     return view('welcome');
 });
 
-Route::get('/new-payment', 'PaymentController@new')->name('payment.new');
-
-Route::post('/new-payment', 'PaymentController@create')->name('payment.create');
-
-Route::get('/payment-callback', 'PaymentController@callback')->name('payment.callback');
+Route::get('/payment', [PaymentController::class, 'create'])->name('payment.create');
+Route::post('/payment', [PaymentController::class, 'store'])->name('payment.store');
+Route::get('/payment/callback', [PaymentController::class, 'callback'])->name('payment.callback');
